requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
websockets==12.0
requests==2.31.0
